Reqorked the magic generator. Optimized for speed
As it turns out the previous, state full implementation suffered quite a
bit in performance. This was due to the way the sate Monad handles
modify. Basically it ends up creating a new instance of the modified
part of the state, and then replaces the old state with a new one. Since
we have Megabyte worth of sate vectors, this is not acceptable.

Now we do purely functional computations to obtain the Magic database.
The code still needs a lot of cleanup though.