Fixed in check captures.
We missed the captures of the checking piece (except when capturer was the same type as the captured). Now we first generate all moves, filter out checks, then regenerate them on path traversal. This needs to be changed.

Initial Perft + some Ruy Lopez that showed the problem more easily now seem to pass perft.