Stateless move + fix depth comparison on TPC hit
Reworked the move generator to be stateless, simplifying the code quite
a bit. Fixed the direction of depth comparison on TPC hits, depth has a
reverse meaning (otherwise you would have to negate it when it goes in
the TPC (aka TT)).

For some reason before I had ~30% hit that is now degraded quite a bit
down to 1-5% depending on the position, and Fine 70 gives only
~25%. (The depth comparison fix doesn't seem to change the percentages.)