Fixed a logic error in simpleChecks
The simple check code had a logic bug that caused non-pawn pieces to
teleport across the board in certain situations. For instance in this
position:

   r7/p1pk3p/1pn3p1/3B4/Q4B2/P4P1P/KPP5/4r3 b - - 0 1

The a8 Rook jumped to a1 to give a check. This was because we had an
other rook attacking a1. The rays casted from the enemy king were and-ed
with the `mconcat`-ed rays casted from all rooks. Therefore we thought
a8a1 was legal move (and it also passed the rook move validity check).