Reworked Zobrist hash calcs with dynamic updates
The zobrist hash was always recalculated from scratch. This was very
expensive with the 64 calls to `pieceAt` for each square. Now the hash
is calculated from the previous hash and the move properties
`xor`ed. The calculation from scratch has been moved into the Board
module where on `fromFEN` we still use it. We could (should?) have a
quickCheck property that the two are equal..