Fix detecting pieces that can do discovered check
The logic to calculate pieces that do discovered checks were broken.

We casted a ray from the enemy king, then from our piece and
anded the 2 with our pieces. This logic is broken because it's possible
that the enemy king sees the found piece on a different ray from the one
that the ray caster / sliding attack piece had.