Removed profiling flag
The --enable-executable-profiling & --enable-library-profiling works
just as well, assuming that we have the TemplateHaskell extension, and
that is the standard way of doing this. Before any of the libraries are
installed, one have to enable library profiling in cabal config as well.

The --auto-all for ghc is still needed, otherwise we just get a bunch of CAFs.