Split up the Killer heuristics to PV // Killer
The implementation was a mixture of PV, killer, and counter move
heuristics that didn't work well, but more to the point it was difficult
to understand what it did.

Now we have a proper Killer move heuristics that is depth and parent
node dependent, and affects only the non-leftmost children of a
node. The store includes a popularity count per entry, which controls
eviction.

The PV heuristics is only applied to the left most path in the tree, it
turns out that stubbornly playing PV moves in totally unrelated position
made more harm than good.