Fix race condition in the handling of the stop command
The stop command handler set the abort, and the polled for the abort
flag to be dropped by the search. This wasn't synchronous with the next
go command. In UCI it seems that after sending the stop we can receive a
go before we responded to the stop. What's the point in any response
then anyways??!! The fix is to move the block in the abort from the end
of the stop handler to the front of the go handler.