Fixes #4. Added time control.
This is a work in progress, and it's far from perfect. There are a lot
of assumptions in the code about how to predict the moves to go, or the
required time the next level of iterative deepening would require.

For 1 minute sudden death time control atm we lose on time, as the
minimal time required to make any move is still too slow. For longer
games the time control already allocates sensible looking times for the
moves.