UCI parser fixes
The parser had a pug with the Go command parsing, we couldn't properly
parse the ponder string at the end of the command after some time
control options. This caused an issue with xboard using polyglot:

   1404035334.763 Adapter->Engine: go wtime 295030 btime 300000 movestogo 39 ponder
   1404035334.764 Engine->Adapter: info string parse error : (line 1, column 43):
   1404035334.764 Adapter->GUI: 1 +0 0 0 parse error : (line 1, column 43):
   1404035334.764 Engine->Adapter: unexpected "p"
   1404035334.765 Engine->Adapter: expecting space, "movetime", "depth", "infinite", "wtime", "btime", "winc", "binc" or "movestogo"

Now this is fixed. Also fixed bad parsec error reporting by reducing the
scope of tries. (try command). A description of this can be found here:
http://blog.ezyang.com/2014/05/parsec-try-a-or-b-considered-harmful/
. This came up with bad FEN strings, reporting bad error messages. The
missing second '-' failes the entire parser, even though we have
established the "position" string at the front. Then the reported error
becomes:

   position fen 4r3/1p1k2p1/2p4p/5p2/Q4p1r/P2P1q1P/1PP2P1K/2R3R1 w - 0 0
   info string parse error : (line 1, column 1):
   unexpected "s"
   expecting "ponderhit"

With the fix the error is the expected:

   info string parse error : (line 1, column 67):
   unexpected "0"
   expecting space or "-"