Defend moves when in check. Delaying move legality.
This is a WIP. The test don't even build.

Reworked - yet again - the move generator. Now when we are in check we
only generate defending moves. Otherwise the legality check has been
delayed after the beta cuts happen. The legality check is different when
we are not in check: we check for pins (unless it's a King move).